Harden local usage history scans
